+++ bitmatops.cpp
@@ -6188,7 +6188,7 @@ unsigned long long int get_size_of_gpu_input(BitMat *bitmat){
 	return size;
 }
 
-void convert_bitmat_to_gpu_input(BitMat* bitmat, unsigned char* gpu_input, long long int* mapping, unsigned int num_subs){
+void convert_bitmat_to_gpu_input(BitMat* bitmat, unsigned char* gpu_input, int* mapping, unsigned int num_subs){
 	unsigned char* curr_row = gpu_input;
 
 	if(bitmat->bm.size() > 0){
